# packages/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AEOS API",
    "description": "Local-first API over the AEOS daemon. Envelope: { success, data, error, meta }.",
    "version": "0.1.0"
  },
...
  "paths": {
...
    "/v1/objectives/{id}/start": {
      "post": {
        "tags": [
          "objectives"
        ],
        "description": "Start (or resume) the objective through the sequential scheduler. Idempotent while running.",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/v1/objectives/{id}": {
      "get": {
        "tags": [
          "objectives"
        ],
        "description": "Objective status derived from plan.md + checkpoints (files are truth).",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/v1/approvals/{requestId}": {
      "post": {
        "tags": [
          "approvals"
        ],
        "description": "Answer a pending approval. Unanswered requests deny on expiry.",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "requestId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    }
  }
}